mainfolder/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django
django-crontab==0.7.1
django-restframework==0.0.1
dnspython==2.6.1
gevent==24.2.1
greenlet==3.0.3
idna==3.6
kombu==5.3.5
logzero==1.7.0
numpy
pandas
prompt-toolkit==3.0.43
pymongo==4.6.2
pyotp==2.9.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
smartapi-python==1.4.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websocket==0.2.1
zope.event==5.0
zope.interface==6.2
gunicorn
